+++ src/main/java/simpleerp/inventory/InventoryTransactionDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record InventoryTransactionDTO(
 
         @NotNull
@@ -12,5 +14,5 @@ public record InventoryTransactionDTO(
         @NotNull
         @Digits(integer =10, fraction=2, message = "Currently supported (10,2)")
         @Min(value = 0, message = "qty cannot be less than 0")
-        double qty
+        BigDecimal qty
 ) {}
