+++ src/main/java/simpleerp/inventory/UpdateInventoryDTO.java
@@ -4,16 +4,18 @@ import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record UpdateInventoryDTO (
 
         @NotNull
         @Digits(integer =10, fraction=2, message = "Currently supported (10,2)")
         @Min(value = 0, message = "qty cannot be less than 0")
-        double qtyAvailable,
+        BigDecimal qtyAvailable,
 
         @NotNull
         @Digits(integer =10, fraction=2, message = "Currently supported (10,2)")
         @Min(value = 0, message = "qty cannot be less than 0")
-        double qtyReserved
+        BigDecimal qtyReserved
 
 ){ }
