+++ src/main/java/simpleerp/reservation/ReservationRequestDTO.java
@@ -1,7 +1,9 @@
 package simpleerp.reservation;
 
+import java.math.BigDecimal;
+
 public record ReservationRequestDTO(
         Long orderId,
         Long componentId,
-        double qty
+        BigDecimal qty
 ) {}
